feat(scripts): add compile time message to output
It was complicated than I thought it would be. Firstly because,
webpack multi compiler always report for all instances. So I had
to store last stat (.toJson, not the actual object, because it
would produce the same result corresponding the current one) and
compare with current one against `builtAt`. If that value changes
then webpack has compiled the instance.

See #444